requirements.txt
datasets
transformers
torch
Pillow
accelerate
sentencepiece
peft
bitsandbytes
matplotlib